Convert all files to Unix line endings
Previously the repo was a mix of Windows and Unix line endings and that
really seemed to confuse some of the tooling I want to use to clean
things up.

Change everything to Unix line endings as that seems to be the
recommended option. You can translate between the two verisons on
checkout by enabling the following git settting:

    git config core.autocrlf true

As this change touched almost every file you lose a lot of information
from git blame. If you have git v2.23 or later all is not lost:

    git blame --ignore-rev HASH_OF_THIS_COMMIT

This will attribute line changes to the previous, more useful, commit.